[CoreBundle] implement identifier php template helper, #260
