+++ src/CoreShop/Bundle/CoreBundle/Twig/CheckoutIdentifierExtension.php
@@ -12,52 +12,21 @@
 
 namespace CoreShop\Bundle\CoreBundle\Twig;
 
-use CoreShop\Component\Order\Checkout\CheckoutManagerFactoryInterface;
-use CoreShop\Component\Order\Checkout\CheckoutManagerInterface;
-use CoreShop\Component\Order\Checkout\ValidationCheckoutStepInterface;
-use CoreShop\Component\Order\Context\CartContextInterface;
-use Symfony\Component\HttpFoundation\RequestStack;
-use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use CoreShop\Component\Order\Model\CartInterface;
+use CoreShop\Bundle\CoreBundle\Templating\Helper\CheckoutIdentifierHelperInterface;
 
 final class CheckoutIdentifierExtension extends \Twig_Extension
 {
     /**
-     * @var RequestStack
+     * @var CheckoutIdentifierHelperInterface
      */
-    private $requestStack;
+    private $helper;
 
     /**
-     * @var UrlGeneratorInterface
+     * @param CheckoutIdentifierHelperInterface $helper
      */
-    protected $router;
-
-    /**
-     * @var CheckoutManagerFactoryInterface
-     */
-    protected $checkoutManagerFactory;
-
-    /**
-     * @var CartContextInterface
-     */
-    protected $cartContext;
-
-    /**
-     * @param RequestStack                    $requestStack
-     * @param UrlGeneratorInterface           $router
-     * @param CheckoutManagerFactoryInterface $checkoutManagerFactory
-     * @param CartContextInterface            $cartContext
-     */
-    public function __construct(
-        RequestStack $requestStack,
-        UrlGeneratorInterface $router,
-        CheckoutManagerFactoryInterface $checkoutManagerFactory,
-        CartContextInterface $cartContext
-    ) {
-        $this->requestStack = $requestStack;
-        $this->router = $router;
-        $this->checkoutManagerFactory = $checkoutManagerFactory;
-        $this->cartContext = $cartContext;
+    public function __construct(CheckoutIdentifierHelperInterface $helper)
+    {
+        $this->helper = $helper;
     }
 
     /**
@@ -66,150 +35,8 @@ final class CheckoutIdentifierExtension extends \Twig_Extension
     public function getFunctions()
     {
         return [
-            new \Twig_Function('coreshop_checkout_steps', [$this, 'getSteps']),
-            new \Twig_Function('coreshop_checkout_steps_*', [$this, 'getStep'])
+            new \Twig_Function('coreshop_checkout_steps', [$this->helper, 'getSteps']),
+            new \Twig_Function('coreshop_checkout_steps_*', [$this->helper, 'getStep'])
         ];
-    }
-
-    /**
-     * Get all Steps of Checkout (cart is always first step here)
-     *
-     * @return array
-     */
-    public function getSteps()
-    {
-        $cart = $this->cartContext->getCart();
-        $request = $this->requestStack->getMasterRequest();
-        $stepIdentifier = $request->get('stepIdentifier');
-        $requestAttributes = $request->attributes;
-        $checkoutManager = $this->checkoutManagerFactory->createCheckoutManager($cart);
-        $currentStep = $checkoutManager->getCurrentStepIndex($stepIdentifier);
-        $checkoutSteps = $checkoutManager->getSteps();
-
-        //always add cart to checkout
-        $isCartStep = $requestAttributes->get('_route') === 'coreshop_cart_summary';
-
-        $shopSteps = [
-            'cart' => [
-                'waiting' => false,
-                'done'    => !is_null($stepIdentifier),
-                'current' => $requestAttributes->get('_route') === 'coreshop_cart_summary',
-                'url'     => $isCartStep
-            ]
-        ];
-
-        foreach ($checkoutSteps as $identifier) {
-            $stepIndex = $checkoutManager->getCurrentStepIndex($identifier);
-            $step = $checkoutManager->getStep($identifier);
-            $isDone = $step instanceof ValidationCheckoutStepInterface ? $step->validate($cart) : false;
-
-            $shopSteps[(string)$identifier] = [
-                'waiting' => is_null($stepIdentifier) || (int)$currentStep < $stepIndex,
-                'done'    => $isDone,
-                'current' => !is_null($stepIdentifier) && (int)$currentStep === $stepIndex,
-                'url'     => !is_null($stepIdentifier) ? $this->router->generate('coreshop_checkout', ['stepIdentifier' => (string)$identifier]) : null
-            ];
-        }
-
-        return $shopSteps;
-    }
-
-    /**
-     * @param string $type
-     * @return mixed
-     */
-    public function getStep($type = '')
-    {
-        $validGuesser = ['get_first', 'get_previous', 'get_current', 'get_next', 'get_last'];
-
-        $getter = lcfirst(str_replace('_', '', ucwords($type, '_'))) . 'Step';
-        if (!method_exists($this, $getter)) {
-            throw new \InvalidArgumentException(sprintf('invalid identifier guess "%s", available guesses are: %s', $type, implode(', ', $validGuesser)));
-        }
-
-        $cart = $this->cartContext->getCart();
-        $checkoutManager = $this->checkoutManagerFactory->createCheckoutManager($cart);
-        $request = $this->requestStack->getMasterRequest();
-        $stepIdentifier = $request->get('stepIdentifier');
-
-        return $this->$getter($cart, $stepIdentifier, $checkoutManager);
-
-    }
-
-    /**
-     * @param CartInterface            $cart
-     * @param string                   $stepIdentifier
-     * @param CheckoutManagerInterface $checkoutManager
-     * @return mixed
-     */
-    protected function getCurrentStep($cart, $stepIdentifier, $checkoutManager)
-    {
-        return $stepIdentifier;
-
-    }
-
-    /**
-     * @param CartInterface            $cart
-     * @param string                   $stepIdentifier
-     * @param CheckoutManagerInterface $checkoutManager
-     * @return mixed
-     */
-    protected function getFirstStep($cart, $stepIdentifier, $checkoutManager)
-    {
-        $steps = $checkoutManager->getSteps();
-        return reset($steps);
-
-    }
-
-    /**
-     * @param CartInterface            $cart
-     * @param string                   $stepIdentifier
-     * @param CheckoutManagerInterface $checkoutManager
-     * @return mixed
-     */
-    protected function getLastStep($cart, $stepIdentifier, $checkoutManager)
-    {
-        $steps = $checkoutManager->getSteps();
-        return end($steps);
-    }
-
-    /**
-     * @param CartInterface            $cart
-     * @param string                   $stepIdentifier
-     * @param CheckoutManagerInterface $checkoutManager
-     * @return mixed
-     */
-    protected function getPreviousStep($cart, $stepIdentifier, $checkoutManager)
-    {
-        $data = null;
-        $request = $this->requestStack->getMasterRequest();
-        $stepIdentifier = $request->get('stepIdentifier');
-        if (!is_null($stepIdentifier)) {
-            if ($checkoutManager->hasPreviousStep($stepIdentifier)) {
-                $step = $checkoutManager->getPreviousStep($stepIdentifier);
-                $data = $step->getIdentifier();
-            }
-        }
-        return $data;
-    }
-
-    /**
-     * @param CartInterface            $cart
-     * @param string                   $stepIdentifier
-     * @param CheckoutManagerInterface $checkoutManager
-     * @return mixed
-     */
-    protected function getNextStep($cart, $stepIdentifier, $checkoutManager)
-    {
-        $data = null;
-        if (!is_null($stepIdentifier)) {
-            if ($checkoutManager->hasNextStep($stepIdentifier)) {
-                $step = $checkoutManager->getNextStep($stepIdentifier);
-                $data = $step->getIdentifier();
-            }
-        }
-
-        return $data;
-
     }
 }
